feat(cat-data-service): Move the panic handling to its own module
